refactor: merge shared/types.ts into shared/wire.ts
TaskStatus was the only definition in shared/types.ts; it now lives in
shared/wire.ts alongside the other shared domain types. Deletes the
now-empty file and updates all imports.

Closes #683
