+++ frontend/src/types.ts
@@ -1,2 +1,1 @@
-export type { TaskStatus } from "../../shared/types.ts";
-export type { Task, Worker, LogEntry, AdminMessage, BlockerInfo } from "../../shared/wire.ts";
+export type { TaskStatus, Task, Worker, LogEntry, AdminMessage, BlockerInfo } from "../../shared/wire.ts";
