--- shared/types.ts
@@ -1,3 +1,0 @@
-// Types shared between the backend (src/) and frontend (frontend/src/).
-
-export type TaskStatus = "pending" | "assigned" | "pushed" | "merged" | "closed" | "complete" | "blocked";
